fix: append correct newline char for fileformat
Problem: the getline() function strips newlines from returned text, so
we must append a newline char to find newlines in visual mode. The
newline character searched for depends on the fileformat, but we
currently just append '\n'. This means that, if the fileformat is not
Linux, the plugin will not currently find newline characters.

Solution: append the correct newline character depending on the
fileformat.