feat: mark whitespace in blockwise visual mode
getregionpos provides information about the cursor moving past the end
of the line, such as with virtualedit = "all". This can be used to
easily determine where to stop highlighting in visual block
mode. Refactor the position-generating function and marking algorithm so
that they get and use this information to provide blockwise highlighting.