fix: keep highlights when switching between visual modes
Problem:
Switching from visual mode to any other mode cleared the highlights.

Solution:
We must have two cases: switching from visual mode to a non-visual mode
which the plugin can highlight in and to a visual mode that the plugin
can highlight in. This way, when we switch from a visual mode that the
plugin supports to another, we keep our highlights. When we switch to
any other mode, we clear them.

Also, rename the aucmd initialization function for clarity.

Fixes #12.