fix: ensure that the setup function runs only once
Problem:
In c42deae9ae48a7991e879408771f8cfbab5961ae, I setup a way to allow
the user to run the plugin without explicitly calling a setup
function. In this commit, I do not give a way for the plugin to check
if the setup function has been called already. It is not idempotent,
and will run twice, registering the plugin's autocommands twice.
This degrades performance.

Solution:
Make the setup function idempotent.

Fixes #22.