fix: use compatible lazy loading event
Problem:
Loading on 'ModeChanged' prevents the plugin from highlighting the
first character on entering visual mode. This happens because the
plugin is loading on entering visual, but it is not activating the
highlighting function in setup.

Solution:
Use a lazy event that is compatible with the existing codebase.
Designing the code around a specific event feels fragile. Instead, we
should use an event that works with the existing codebase and doesn't
enforce building code to match. If users want to use a different event,
that should be up to them but not up to us to cater to.