fix: tabs are miscalculated
Problem:
getregionpos provides an offset in the fourth position. I forgot that
this also applied to tabs. I don't exactly remember why I was using this
block of code, but it breaks how tab highlighting is calculated, which
is a very important feature.

Solution:
Calculate tabs correctly by not considering the offset value for each
position returned by getregionpos.