fix: fix enabled flag being ignored when set to false
PROBLEM:
When the user sets the `enabled` flag to false in their configuration,
it is ignored. The way enabling was set up, it automatically flipped
the flag to true if the user entered a buffer or filetype that was
not blacklisted.

SOLUTION:
Establish two different types of enabling: user-defined and ft/bt. The
way this works now is that the user-defined enabling is always respected
and the ft/bt enabling is considered only when the user has the
user-defined enabling set to `true`. This also allows for lazy-loading,
incidentally.

Fixes #33
Fixes #29