feat!: define line endings by fileformat
Specify line ending chars in the configuration based on the
file format instead of the character type. This allows us to quickly
determine the fileformat based only on line ending chars that we
highlight.

Internally, Neovim represents all line endings as line feeds. The
mac and dos fileformats do not have their line endings actually
put into the buffer. For us, this means that we never actually
need to search for them in lines we want to highlight. We will
always find line feed characters instead.

Given this, we can simply search for the end of the line and place
whatever character there that we want.