feat: support leading and trailing spaces
To support more powerful patterns, the plugin moves from Lua's
pattern matching to Vim's regular expressions. From there, we define
what patterns we would like to capture. This allows for leading and
trailing spaces, which are implemented, as well as for other patterns
to be supported in the future.

Because there is now a larger incentive to support the user's listchars
option as a default, this has also been implemented. The default
configuration has changed, but the plugin's current behavior is the
same. The new configuration is given in the readme. The only
addition is the "match_types" setting, which tells the plugin what
types of white space to match on. The "list_chars" setting has
changed the default tab character and has added leading and
trailing space characters.

We also now use Stylua to format the source code. A configuration
file is now included with the repo and the according formatting
changes have been made.

Closes #5
Closes #34